Revising the position tracking to go by boundaries
This is rather than width and height. Width and height would be more
useful if we were planning on moving objects around on screen, but since
they'll be pretty much fixed in place it's probably a better use of our
time and computational energy to just be aware of the width and height
during generation.